requirements.txt
absl-py==0.13.0
anyio==3.3.4
appdirs==1.4.4
astunparse==1.6.3
audioread==2.1.9
autopep8 @ file:///tmp/build/80754af9/autopep8_1620866417880/work
backcall==0.2.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.6
clang==5.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
decorator==5.1.0
dominate==2.6.0
ffmpeg-python==0.2.0
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.1
h11==0.12.0
h2==4.1.0
h5py==3.1.0
hpack==4.0.0
httpcore==0.13.7
httpx==0.19.0
hyperframe==6.0.1
idna==3.2
importlib-metadata==4.8.2
ipython==7.27.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.1.0
keras==2.7.0
keras-nightly==2.8.0.dev2021120708
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
librosa==0.8.1
llvmlite==0.36.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
norbert==0.2.1
numba==0.53.1
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.2
pandas==1.3.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.2
pooch==1.5.2
prompt-toolkit==3.0.20
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycodestyle @ file:///tmp/build/80754af9/pycodestyle_1636635402688/work
pycparser==2.20
pydub==0.25.1
Pygments==2.10.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytube==11.0.1
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
resampy==0.2.2
rfc3986==1.5.0
rsa==4.7.2
scikit-learn==1.0.1
scipy==1.7.1
seaborn==0.11.2
six==1.15.0
sniffio==1.2.0
SoundFile==0.10.3.post1
tb-nightly==2.8.0a20211205
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.7.0
tensorflow-gpu==2.7.0
tensorflow-io==0.22.0
tensorflow-io-gcs-filesystem==0.21.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021120709
tf-nightly==2.8.0.dev20211207
threadpoolctl==3.0.0
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tqdm==4.62.3
traitlets==5.1.0
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.6
visitor==0.1.3
wcwidth==0.2.5
Werkzeug==2.0.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.6.0
